moved Some_val to kamlo for reusability
