+++ lablqml/stubs/startup_stubs.c
@@ -11,15 +11,6 @@
 #include <QtQuick/QQuickWindow>
 #include <QtQuick/QQuickItem>
 
-static value Val_some(value v) {
-  CAMLparam1(v);
-  CAMLlocal1(some);
-
-  some = caml_alloc(1, 0);
-  Store_field(some, 0, v);
-
-  CAMLreturn(some);
-}
 /* This is a crazy macro to allocation argc and argv and pass them to
  * QApplication. The problem is that we need copying of argv because they are
  * ocaml values that can be moved. Moreover, we need to allocate argc because
